requirements.txt
annotated-types==0.7.0
anyio==4.8.0
beautifulsoup4==4.13.3
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
fastapi==0.115.11
h11==0.14.0
idna==3.10
pydantic==2.10.6
pydantic_core==2.27.2
python-dotenv==1.0.1
redis==5.2.1
requests==2.32.3
sniffio==1.3.1
soupsieve==2.6
starlette==0.46.0
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
